requirements.txt
python-chess
cryptography